Simplify user switching in postgresql11-install.service
This service needs to first create the data directory for postgresql11,
then run it to create the initial database. The directory must be
created by root, but the database should be created by the postgres
user. We can use the User= directive to make the database creation
command run as the correct user, and the PermissionsStartOnly= directive
to ensure that the directory creation command runs as root beforehand.

This eliminates the need for su and bash in the main ExecStart=
directive.

Fixes https://github.com/clearlinux/distribution/issues/1335